% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate_boundary}
\alias{animate_boundary}
\title{Animate the evolution of a decision boundary as the sample size grows}
\usage{
animate_boundary(sample, density, delta, fit_and_predict, start = 0)
}
\arguments{
\item{sample}{`data.frame`: the complete sample data; should have
columns `x`, `y`, and `class`}

\item{density}{`data.frame`: the density distribution of `x` and
`y`; should have columns `x`, `y`, and `optimal`, the contours
of the optimal decision distribution}

\item{delta}{`integer`: how many points to add at each step of the
animation}

\item{fit_and_predict}{`function(sample, density, ...)`: fits a learner
to the sample data and returns its predictions on the density; `...`
are parameters passed to the learner}
}
\description{
Animate the evolution of a decision boundary as the sample size grows
}
