% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_and_predict.R
\name{fit_and_predict}
\alias{fit_and_predict}
\alias{fit_and_predict_nb}
\alias{fit_and_predict_qda}
\alias{fit_and_predict_mda}
\alias{fit_and_predict_glm}
\alias{fit_and_predict_gam}
\alias{fit_and_predict_mars}
\alias{fit_and_predict_polymars}
\alias{fit_and_predict_knn}
\alias{fit_and_predict_svm}
\alias{fit_and_predict_rpart}
\alias{fit_and_predict_rf}
\alias{fit_and_predict_etrees}
\alias{fit_and_predict_xgboost}
\alias{fit_and_predict_mb}
\alias{fit_and_predict_nn}
\alias{fit_and_predict_elm}
\alias{fit_and_predict_gp}
\title{Fit a model to a sample and predict on a density}
\usage{
fit_and_predict_nb(sample, density, ...)

fit_and_predict_qda(sample, density, ...)

fit_and_predict_mda(sample, density, ...)

fit_and_predict_glm(sample, density, ...)

fit_and_predict_gam(sample, density, ...)

fit_and_predict_mars(sample, density, ...)

fit_and_predict_polymars(sample, density, ...)

fit_and_predict_knn(sample, density, ...)

fit_and_predict_svm(sample, density, ...)

fit_and_predict_rpart(sample, density, ...)

fit_and_predict_rf(sample, density, ...)

fit_and_predict_etrees(sample, density, ...)

fit_and_predict_xgboost(sample, density, ...)

fit_and_predict_mb(sample, density, ...)

fit_and_predict_nn(sample, density, seed = 31415)

fit_and_predict_elm(sample, density, ...)

fit_and_predict_gp(sample, density, ...)
}
\arguments{
\item{sample}{`data.frame`: a sample from a probability
distribution}

\item{density}{`data.frame`: a grid of data points with density
values}

\item{...:}{parameters to pass on to the model}
}
\value{
the density data frame with a column of fitted probability
    predictions from the model
}
\description{
Fit a model to a sample and predict on a density
}
