% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gg_density}
\alias{gg_density}
\title{Make a density layer}
\usage{
gg_density(data, z, size = 1, color = "black", alpha = 1, ...)
}
\arguments{
\item{data}{character: the name of the contour column}
}
\description{
Make a density layer
}
