% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gg_mix_label}
\alias{gg_mix_label}
\title{Make a layer of component labels for a mixture distribution with two classes}
\usage{
gg_mix_label(mus, classes = NULL, size = 10, ...)
}
\arguments{
\item{mus}{list(data.frame): the means for components of each class; every row
is a mean, each column is a coordinate}

\item{classes}{(optional) a vector of which levels of class to plot}
}
\description{
Make a layer of component labels for a mixture distribution with two classes
}
