% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gg_optimal}
\alias{gg_optimal}
\title{Make an optimal boundary layer}
\usage{
gg_optimal(data, breaks = c(0), ...)
}
\arguments{
\item{data}{data.frame: a data grid of features `x` and `y` with a column with
the `optimal` boundary contours}

\item{breaks}{numeric: which contour levels of `optimal` to plot}
}
\description{
Make an optimal boundary layer
}
