% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gg_plot_boundary}
\alias{gg_plot_boundary}
\title{Plot an optimal boundary together with density contours and a sample}
\usage{
gg_plot_boundary(sample, density, title = "")
}
\description{
Plot an optimal boundary together with density contours and a sample
}
