% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gg_sample}
\alias{gg_sample}
\title{Make a sample layer}
\usage{
gg_sample(data, classes = NULL, size = 3, alpha = 0.5, ...)
}
\arguments{
\item{data}{data.frame: a sample with continuous features `x` and `y`
grouped by factor `class`}

\item{classes}{(optional) a vector of which levels of `class` to
plot; default is to plot data from all classes}
}
\description{
Make a sample layer
}
