% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer_density}
\alias{layer_density}
\title{Make a density layer}
\usage{
layer_density(data, z, size = 1, color = "black", alpha = 1, ...)
}
\arguments{
\item{data}{data.frame: a data grid of features `x` and `y` with
contours `z`}

\item{z}{character: the name of the contour column}
}
\value{
a `ggplot2::LayerInstance` object for use with `ggplot2::gg_plot()`
}
\description{
Make a density layer
}
