% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer_sample}
\alias{layer_sample}
\title{Make a sample layer}
\usage{
layer_sample(data, classes = NULL, alpha = 0.5, size = 3, ...)
}
\arguments{
\item{data}{data.frame: a sample with continuous features `x` and
`y` grouped by factor `class`}

\item{classes}{(optional) a vector of which levels of `class` to
plot; default is to plot data from all classes}
}
\value{
a `ggplot2::LayerInstance` object
}
\description{
Make a sample layer
}
