% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{make_density_mix}
\alias{make_density_mix}
\title{Construct a dataframe with posterior class probabilities and the
optimal decision boundary over a grid on the feature space}
\usage{
make_density_mix(
  mean_0,
  sigma_0,
  w_0,
  mean_1,
  sigma_1,
  w_1,
  p_0,
  x_min,
  x_max,
  y_min,
  y_max,
  delta = 0.05
)
}
\arguments{
\item{mean_0}{numeric: the average mean of the components of the first feature}

\item{sigma_0}{matrix: covariance of components of the first feature}

\item{w_0}{numeric: vector of weights for components of the first feature}

\item{mean_1}{numeric: the average mean of the components of the second feature}

\item{sigma_1}{matrix: covariance of components of the second feature}

\item{w_1}{numeric: vector of weights for components of the second feature}

\item{p_0}{double: the prior probability of class 0}
}
\description{
Construct a dataframe with posterior class probabilities and the
optimal decision boundary over a grid on the feature space
}
