% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{make_density_mvn}
\alias{make_density_mvn}
\title{Construct a dataframe with posterior class probabilities and the
optimal decision boundary over a grid on the feature space}
\usage{
make_density_mvn(
  mean_0,
  sigma_0,
  mean_1,
  sigma_1,
  p_0,
  x_min,
  x_max,
  y_min,
  y_max,
  delta = 0.05
)
}
\arguments{
\item{mean_0}{vector: the mean vector of the first class}

\item{sigma_0}{matrix: the 2x2 covariance matrix of the first class}

\item{mean_1}{vector: the mean vector of the second class}

\item{sigma_1}{matrix: the 2x2 covariance matrix of the second class}

\item{p_0}{double: the prior probability of class 0}
}
\description{
Construct a dataframe with posterior class probabilities and the
optimal decision boundary over a grid on the feature space
}
