% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{make_mix_sample}
\alias{make_mix_sample}
\title{Generate normally distributed feature samples for a binary
classification problem}
\usage{
make_mix_sample(
  n,
  nu_0,
  tau_0,
  n_0,
  sigma_0,
  w_0,
  nu_1,
  tau_1,
  n_1,
  sigma_1,
  w_1,
  p_0
)
}
\arguments{
\item{n}{integer: the size of the sample}

\item{nu_0}{numeric: the average mean of the components of the first feature}

\item{n_0}{integer: class frequency of first feature in the sample}

\item{sigma_0}{matrix: covariance of components of the first feature}

\item{w_0}{numeric: vector of weights for components of the first feature}

\item{n_1}{integer: class frequency of second feature in the sample}

\item{sigma_1}{matrix: covariance of components of the second feature}

\item{w_1}{numeric: vector of weights for components of the second feature}

\item{p_0}{double: the prior probability of class 0}

\item{mean_1}{numeric: the average mean of the components of the second feature}
}
\description{
Generate normally distributed feature samples for a binary
classification problem
}
