% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{make_mvn_sample}
\alias{make_mvn_sample}
\title{Generate normally distributed feature samples for a binary
classification problem}
\usage{
make_mvn_sample(n, mu_0, sigma_0, mu_1, sigma_1, p_0)
}
\arguments{
\item{n}{integer: the size of the sample}

\item{sigma_0}{matrix: the 2x2 covariance matrix of the first class}

\item{sigma_1}{matrix: the 2x2 covariance matrix of the second class}

\item{p_0}{double: the prior probability of class 0}

\item{mean_0}{vector: the mean vector of the first class}

\item{mean_1}{vector: the mean vector of the second class}
}
\description{
Generate normally distributed feature samples for a binary
classification problem
}
