% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{new_density_mvn}
\alias{new_density_mvn}
\title{Construct a data frame with posterior class probabilities and the
optimal decision boundary over a grid on the feature space}
\usage{
new_density_mvn(
  mu_0,
  sigma_0,
  mu_1,
  sigma_1,
  p_0,
  x_min,
  x_max,
  y_min,
  y_max,
  delta = 0.05
)
}
\arguments{
\item{mu_0}{vector: the mean vector of the first class}

\item{sigma_0}{matrix: the 2x2 covariance matrix of the first class}

\item{mu_1}{vector: the mean vector of the second class}

\item{sigma_1}{matrix: the 2x2 covariance matrix of the second class}

\item{p_0}{double: the prior probability of class 0}
}
\value{
a DensityFrame object
}
\description{
Construct a data frame with posterior class probabilities and the
optimal decision boundary over a grid on the feature space
}
