% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnimationFrame.R
\name{new_DensityFrameSequence}
\alias{new_DensityFrameSequence}
\title{Create a frame for animating a distribution parameter}
\usage{
new_DensityFrameSequence(density_fun, params, method = expand.grid, ...)
}
\arguments{
\item{density_fun}{`function(...)`: a function returning a
DensityFrame object. The parameters of the function will be
drawn from `params`}

\item{params}{`list`: a named list of parameters with each entry a
list of values for that parameter}

\item{method}{`function`}

\item{...}{: options to be passed to `layer_density` when plotting}
}
\value{
a DensityFrameSequence object
}
\description{
An internal function.
}
