% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnimationFrame.R
\name{new_FittedFrameSequence}
\alias{new_FittedFrameSequence}
\title{Make an animation frame for a "fitting" animation}
\usage{
new_FittedFrameSequence(
  sample_frame,
  density_frame,
  fit_and_predict,
  delta = 20L,
  start = 1L,
  ...
)
}
\arguments{
\item{sample_frame}{a sample frame}

\item{density_frame}{a density frame}

\item{fit_and_predict}{prediction function}

\item{...}{}
}
\value{
a FittedFrameSequence
}
\description{
Make an animation frame for a "fitting" animation
}
