% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnimationFrame.R
\name{new_model_frame}
\alias{new_model_frame}
\title{Make an animation frame for a "model" animation}
\usage{
new_model_frame(sample_frame, density_frame, fit_and_predict, ...)
}
\description{
Make an animation frame for a "model" animation
}
