% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnimationFrame.R
\name{new_sample_frame}
\alias{new_sample_frame}
\title{Make an animation frame for a "sample" animation}
\usage{
new_sample_frame(sample_fun, ...)
}
\description{
Make an animation frame for a "sample" animation
}
