% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{optimal_predict}
\alias{optimal_predict}
\title{Make an optimal prediction at a point from two class distributions}
\usage{
optimal_predict(xy, p_0, dfun_0, dfun_1)
}
\arguments{
\item{xy}{vector: 2D coordinate}

\item{p_0}{double: prior probability of class 0}

\item{dfun_0}{function(x): density of features of class 0}

\item{dfun_1}{function(x): density of features of class 1}
}
\description{
Make an optimal prediction at a point from two class distributions
}
