% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{sequence_data_parameter}
\alias{sequence_data_parameter}
\title{Create a data frame sequenced for animation of a model parameter}
\usage{
sequence_data_parameter(data, fit_and_predict, params)
}
\arguments{
\item{data}{`data.frame`: the data to sequence}

\item{fit_and_predict}{`function(sample, density, param)`: fits a
learner to the sample data and returns its predictions on the
density as a data frame; `param` is the parameter value to
vary over}

\item{params}{`list`: a named list of parameters with each entry a
vector of values for that parameter}
}
\value{
a data frame with `n` successive groups of `data`
}
\description{
A function unsed internally by animate_model_parameter
}
