% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{sequence_data_sample}
\alias{sequence_data_sample}
\title{Create a data frame sequenced for animation of a sample}
\usage{
sequence_data_sample(data, delta = 1L)
}
\arguments{
\item{data}{`data.frame`: a data frame to sequence}

\item{delta}{`int`: how many rows to increase by at each iteration}
}
\value{
a data.frame with successive groups of `data` of size `delta`;
`data` is randomized before sequencing
}
\description{
Create a data frame sequenced for animation of a sample
}
